requirements.txt
Django==3.1.7
djangorestframework~=3.12.2
requests~=2.25.1
payping~=0.0.4
easy_thumbnails~=2.7.1